Move test cases to PolarisServiceImplIntegrationTest
